tee_rpmb.c: allocate two payload buffers for request and response
It is not valid to allocate a single buffer with
thread_optee_rpc_alloc_payload(), then split it into two parameters of
type TEESMC_ATTR_TYPE_MEMREF_* in teesmc32_param. The linux driver
would not be able to map the second buffer.
Thus, this commit replaces the single allocation with two smaller ones.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>