Travis: cache toolchains [Not useful currently]
Download and extract the toolchain packages into the cached download
directory so that they are readily available when the build starts.
Avoids needlessly accessing the Linaro servers.

NOTE: Currently does *NOT* significantly reduce the build time, it is
possibly worse because:
- download/extract time is about the same. Linaro servers are on
Amazon S3 and DL is fast, the Travis cache file is also on S3 so no
difference here.
- cache packing/upload is longer.
This should be tried again if/when the cache packing/upload occurs
asynchronously [1].

[1] https://docs.travis-ci.com/user/caching/ "The upload is currently
part of the build cycle, but we’re looking into improving that to
happen outside of the build, giving faster build feedback."

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>