core: storage: close file handle before removing object
I've hit an interesting issue when running the secure storage tests on
HiSilicon D02. The problem is with REE FS, when the data directory is
mounted over NFS. Running xtest 6009 followed by 6010 fails:
TEE_StartPersistentObjectEnumerator() returns TEE_SUCCESS instead of
TEE_ERROR_ITEM_NOT_FOUND.

This is because test 6009 does not properly clean up. The test creates
persistent objects, closes them, opens them again with the meta accesss
flag and deletes them by calling TEE_CloseAndDeletePersistentObject1().
The secure storage is supposed to be in its initial state on exit.
Instead, when NFS is used, the TA directory
(/data/tee/A7F289B6DF8A7A479F9932E90C0AD0A2) is not deleted -- it is
left empty. The enumerator code called by test 6010 does not expect
that and fails.

This commit fixes TEE_CloseAndDeletePersistentObject1().

It turns out that tee-supplicant *does* attempt to remove the directory
but receives error ENOTEMPTY when the filesystem is NFS-mounted. The
reason is that when it removes the file, as a consequence of
syscall_storage_obj_del() calling o->pobj->fops->remove(o->pobj),
tee-supplicant still has an open file descriptor to the object. This
causes the NFS server to keep a hidden file (.nfsXXX) in the directory,
causing the subsequent rmdir() to fail. Then when
syscall_storage_obj_del() calls tee_obj_close(), the fd is closed and
the .nfsXXX file disappears. We're left with an empty TA directory.

Fix that by closing the file handle first. A helper function is
introduced since there are several call sites.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Tested-by: Jerome Forissier <jerome.forissier@linaro.org> (D02, NFS, GP)
Tested-by: Jerome Forissier <jerome.forissier@linaro.org> (QEMU)
Tested-by: Jerome Forissier <jerome.forissier@linaro.org> (HiKey, GP)