Add support for early Trusted Applications
Early TAs are user-mode Trusted Applications that are embedded at link
time in the TEE binary. A special read-only data section is used to
store them (.rodata.early_ta). A Python script takes care of converting
the TAs into a C source file with the proper linker section attribute.

The feature is disabled by default. To enable it, the paths to the TA
binaries have to be given in $(EARLY_TA). They should be ELF files.
Typical build steps:
  $ make ... CFG_EARLY_TA=y                         # (1)
  $ # ... build the TAs ...                         # (2)
  $ make ... EARLY_TA=path/to/<uuid>.stripped.elf   # (3)
Notes:
- Setting CFG_EARLY_TA=y during the first step (1) is not necessary,
   but it will avoid rebuilding during the third step (3)
- CFG_EARLY_TA is automatically enabled when EARLY_TA is non-empty in
  step (3)
- Several TAs may be given in $(EARLY_TA) (3)

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>