Add support for several user TA stores
Replace the pointer to the user-mode TA load operations with a list
so that several implementations may be used simultaneously.
Make tee_ta_init_user_ta_session() iterate on the list and stop as
soon as a matching TA is found.
This is preparatory work for the introduction of a new TA store.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>