core/lib/miniz: mz_inflate(): allow NULL pointer for next_out
It can happen that one wants to decompress data but discard parts of
the output. A use case is loading a compressed ELF file; some sections
may not be relevant to the load operation and should therefore be
ignored.
Unfortunately, mz_inflate() does not allow pStream->next_out to be
NULL. As a result, the caller has to allocate a buffer to store the
data, which can be big, only to discard it later.
This patch modifies mz_inflate() so that data is not copied, and
next_out is not incremented, when next_out is NULL when entering the
function. Thanks to this change, the caller who wants to skip n bytes
of decompressed data just needs to set pStream->avail_out = n and
pStream->next_out = NULL.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>