Add support for compressed early TAs
Add decompression code to the early TA loader and update the Python
script accordingly. The compression algorithm is DEFLATE, which is used
by zlib and gzip in particular.
Tested on QEMU (armv7), the compression ratio was between 3 (for bigger
TAs) and 4.7 (for smaller ones). The code size overhead for the loader,
including the decompressor, is 8K with DEBUG=0 or 16K with DEBUG=1. The
decompressor allocates about 43K of heap.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>